Fixup the view-mode for recent changes made in webform.
webform-7.x-3.23 and 7.x-4.5 introduced a change in when webforms are
rendered. This leads to the block not being shown if the webform is not
shown in the full-view of the node. In addition to this commit those
users need to apply the patch from [1] to webform or wait until the next
webform-release: webform-7.x-3.24 and webform-7.x-4.7.

[1] https://www.drupal.org/node/1919872#comment-9754913